fix: validate.js
